requirements.txt
GitPython==0.3.2.RC1
async==0.6.1
docopt==0.6.1
gitdb==0.5.4
smmap==0.8.2
beautifulsoup4>=4.3.2
lxml>=3.4.1